refactor(directory selection): refactor directory selection GUI code
Separate business logic from UI logic
Separate "New project" into a different sub-application window
  - This simplifies the UI making it clearer for newbie users
  - Makes it easier to extend "New Project"
Allows "open project" to be more easily extended in the future to include multiple last recent projects
Increases the testability a lot, coverage should well above 84% now
