+++ ardupilot_methodic_configurator/frontend_tkinter_parameter_editor.py
@@ -223,13 +223,11 @@ class ParameterEditorWindow(BaseWindow):  # pylint: disable=too-many-instance-at
 
         # Create a new frame inside the config_subframe for the intermediate parameter file directory selection labels
         # and directory selection button
-        connected_vehicle_type = self.configuration_manager.connected_vehicle_type
         directory_selection_frame = VehicleDirectorySelectionWidgets(
             self,
             config_subframe,
             self.local_filesystem.vehicle_dir,
             destroy_parent_on_open=False,
-            connected_fc_vehicle_type=connected_vehicle_type,
         )
         if self.gui_complexity != "simple":
             directory_selection_frame.container_frame.pack(side=tk.LEFT, fill="x", expand=False, padx=(4, 6))
